charsets.c / markdown.cpp / stringdiffs.cpp: Fix 64 bit build warnings

+++ Src/charsets.c
@@ -986,7 +986,7 @@ static int SortCompareByCodePage(const void *elem1, const void *elem2)
 {
 	const struct _charsetInfo *p = *(const struct _charsetInfo **)elem1;
 	const struct _charsetInfo *q = *(const struct _charsetInfo **)elem2;
-	return p->codepage != q->codepage ? p->codepage - q->codepage : p - q;
+	return p->codepage != q->codepage ? p->codepage - q->codepage : (int)(p - q);
 }
 
 static struct _charsetInfo const *FindByName(const char *name)
